pyproject.toml
[project]
name = "course"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "robotframework>=7.3.2",
    "robotframework-browser[bb]>=19.9.0",
    "robotframework-databaselibrary>=2.3.2",
    "robotframework-pabot>=5.1.0",
    "robotframework-requests>=0.9.7",
    "robotframework-robocop>=6.8.0",
    "robotframework-seleniumlibrary>=6.8.0",
    "ruff>=0.14.0",
]
